refactor(register): step 2, cleanup props, types, selectors, rename components and organize
