+++ packages/blockchain-wallet-v4-frontend/src/scenes/Signup/components/Header/index.tsx
