+++ packages/blockchain-wallet-v4-frontend/src/scenes/app.tsx
@@ -30,7 +30,7 @@ const Logout = React.lazy(() => import('./Logout'))
 const MobileLogin = React.lazy(() => import('./MobileLogin'))
 const RecoverWallet = React.lazy(() => import('./RecoverWallet'))
 const RecoverWalletLegacy = React.lazy(() => import('./RecoverWalletLegacy'))
-const Register = React.lazy(() => import('./Register'))
+const Signup = React.lazy(() => import('./Signup'))
 const ResetWallet2fa = React.lazy(() => import('./ResetWallet2fa'))
 const ResetWallet2faToken = React.lazy(() => import('./ResetWallet2faToken'))
 const UploadDocuments = React.lazy(() => import('./UploadDocuments'))
@@ -91,7 +91,7 @@ const App = ({
                     />
                     <PublicLayout path='/reset-2fa' component={ResetWallet2fa} />
                     <PublicLayout path='/reset-two-factor' component={ResetWallet2faToken} />
-                    <PublicLayout path='/signup' component={Register} />
+                    <PublicLayout path='/signup' component={Signup} />
                     <PublicLayout path='/verify-email' component={VerifyEmailToken} />
                     <PublicLayout
                       path='/upload-document/success'
